Narrow opcode factory validation and isolate weaver tests
